move new NSRange helpers into dedicated files
