+++ Sources/DeclarativeTextKit/NSRange+expandedToOther.swift
@@ -1,0 +1,28 @@
+//  Copyright © 2024 Christian Tietze. All rights reserved. Distributed under the MIT License.
+
+import Foundation
+
+extension NSRange {
+    @inlinable
+    func expanded(
+        to other: NSRange,
+        direction: Direction
+    ) -> NSRange {
+        precondition(other.location <= self.location && other.endLocation >= self.endLocation, "Expansion requires other range to be larger")
+
+        let startLocation = switch direction {
+        case .upstream: other.location
+        case .downstream: self.location
+        }
+
+        let endLocation = switch direction {
+        case .upstream: self.endLocation
+        case .downstream: other.endLocation
+        }
+
+        return NSRange(
+            startLocation: startLocation,
+            endLocation: endLocation
+        )
+    }
+}
